Restructure codebase into core/, input/, and audio/ submodules
Major reorganization of the three largest files (`app.rs`, `main.rs`, `pipe.rs`)
into focused, single-responsibility modules. No behavioral changes and all
functionality is preserved, all existing tests pass.

== New module: `core/` ==

Extracted from `app.rs` (1083 → 10 lines, now a thin re-export facade):

- `core/types.rs`: `InputMode`, `ActivePanel`, `Overlay`, `QueryKind`, `NowPlaying`,
  `SongLogEntry` are domain types that were mixed in with UI state
- `core/perf.rs`: `PerfStats`, `FrameTiming`, `PerfSummary` for performance
  profiler infrastructure (~163 lines)
- `core/radio.rs`: `FetchResult`, `fetch_blended_by_tag/name`, `fetch_more`,
  `filter_spam`, `interleave` are RadioBrowser API functions (~218 lines)
- `core/app.rs`: App struct and all its methods for what remains after
  extracting the above (~635 lines)

The original `src/app.rs` is now a facade that re-exports everything from
`core/*`, so all existing `use crate::app::{App, Overlay, ...}` imports
throughout the UI layer continue to work with zero changes.

== New module: `input/` ==

Extracted from `main.rs` (455 → 183 lines):

- `input/handler.rs`: `handle_key()`, `handle_normal()`, `handle_editing()`,
  `handle_theme_picker()`, `handle_genre_picker()`, `handle_settings()` for
  all keybinding dispatch that was inline in the event loop (~313 lines)

`main.rs` is now a clean event loop: terminal setup, draw, poll events,
delegate to input::handler, tick, record perf, teardown.

== Split: `audio/` ==

Extracted from `pipe.rs` (570 → 186 lines):

- `audio/seqlock.rs`: Generic SeqLock<T: Copy> with safety docs — a
  reusable lock-free primitive (~72 lines)
- `audio/fft.rs`: fft_in_place, compute_band_edges, group_into_bands,
  compute_band_energies, plus the full FFT test suite (~313 lines)
- audio/pipe.rs: AudioAnalysis, SharedAnalysis, FIFO creation,
  spawn_reader, reader_loop to just the PCM pipeline (~186 lines)

== Visualizer rendering fix ==

Updated `ui/visualizer.rs` to use proportional per-bar positioning
(i * panel_width / num_bars) instead of uniform stride with integer
division. This eliminates wasted space from rounding on wide panels,
particularly noticeable on 4K displays where 16 bars into ~80 columns
would previously lose ~14 columns to floor division.

== Summary ==

Before → After (lines):
  app.rs:  1083 → 10 (facade)
  main.rs:  455 → 183
  pipe.rs:  570 → 186

No changes to: ui/, storage/, audio/player.rs, audio/visualizer.rs

Closes #29
