+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "AetherTune"
-version = "0.8.2"
+version = "0.9.0"
 edition = "2024"
 
 [dependencies]
