+++ src/audio/mod.rs
@@ -1,3 +1,5 @@
+pub mod seqlock;
+pub mod fft;
 pub mod pipe;
 pub mod player;
 pub mod visualizer;
